requirements.txt
python-docx==0.8.11
regex==2023.10.3
python-dateutil==2.8.2 